% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_ipropath_results.R
\name{assemble_ipropath_results}
\alias{assemble_ipropath_results}
\title{Assemble final iProPath results table}
\usage{
assemble_ipropath_results(
  pairwise_res,
  y_transreg,
  ipropath_res_p,
  adjust_methods = c("BH", "bonferroni", "BY")
)
}
\arguments{
\item{pairwise_res}{Data frame of pairwise regression results.}

\item{y_transreg}{Data frame returned by \code{get_Y_transRegDat()}.}

\item{ipropath_res_p}{Data frame \code{res_p} from \code{ipropath()} output.}

\item{adjust_methods}{Character vector specifying multiple-testing corrections for iProPath p-values.}
}
\value{
A data frame containing merged pairwise and iProPath results.
}
\description{
Assemble final iProPath results table
}
