% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pi_df.R
\name{compute_pi_df}
\alias{compute_pi_df}
\title{Compute pi dataframe}
\usage{
compute_pi_df(pi1_MX, pi1_YM)
}
\arguments{
\item{pi1_MX}{estimated proportion of non-null signals for M ~ X p-values.}

\item{pi1_YM}{estimated proportion of non-null signals for Y ~ M p-values.}
}
\value{
data frame
}
\description{
Compute pi dataframe
}
