% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.time_block}
\alias{.time_block}
\title{Time a code block and print elapsed time}
\usage{
.time_block(label, expr)
}
\arguments{
\item{label}{A short character string used as the message label.}

\item{expr}{An expression to be evaluated.}
}
\value{
The result of evaluating \code{expr}.
}
\description{
A lightweight helper to measure and report the running time of an
expression using \code{message()}. Mainly intended for internal
progress/diagnostic messages in package functions.
}
\keyword{internal}
