% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_data_g}
\alias{extract_data_g}
\title{Extract rows matching a given key from a list of data frames}
\usage{
extract_data_g(df_list, match_key_g)
}
\arguments{
\item{df_list}{list of data frames}

\item{match_key_g}{value of matching_key to extract}
}
\value{
data frame
}
\description{
Extract rows matching a given key from a list of data frames
}
