% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Y_transRegDat.R
\name{get_Y_transRegDat}
\alias{get_Y_transRegDat}
\title{Build outcome regression data for a given X and its cis-mediators M}
\usage{
get_Y_transRegDat(
  x_df_list,
  fixed_cov,
  y_covdf_list,
  mediator_list,
  Y_list,
  x_name,
  y_name,
  num_cores
)
}
\arguments{
\item{x_df_list}{named list of data frames for X}

\item{fixed_cov}{data frame of fixed covariates (optional, depends on your pipeline)}

\item{y_covdf_list}{data frame of covariates associated with Y (optional, depends on your pipeline)}

\item{mediator_list}{list of mediator data frames (each must include matching_key)}

\item{Y_list}{named list of outcome (Y) data frames.}

\item{x_name}{character, which element in x_df_list to use}

\item{y_name}{character, name of the outcome in Y_list}

\item{num_cores}{integer, number of cores for parallel backend}
}
\value{
data frame (row-bind results across matching keys)
}
\description{
Build outcome regression data for a given X and its cis-mediators M
}
