% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipropath.R
\name{ipropath}
\alias{ipropath}
\title{Integrative pathway-based mediation analysis (iProPath}
\usage{
ipropath(
  p_value_YX,
  p_value_MX,
  p_value_YM,
  pi1_MX,
  pi1_YM,
  idx_Xgene,
  idx_Mediator,
  idx_Ygene,
  idx_Ysite,
  Est_param_YX,
  Est_param_YM,
  Est_param_MX,
  numCores = 1,
  y_name
)
}
\arguments{
\item{p_value_YX}{numeric vector. P-values from get_Y_transRegDat.R}

\item{p_value_MX}{Numeric vector. P-values from get_M_RegDat.R}

\item{p_value_YM}{Numeric vector. P-values from get_Y_transRegDat.R}

\item{pi1_MX}{estimated proportion of non-null signals for \eqn{M \sim X} p-values.}

\item{pi1_YM}{estimated proportion of non-null signals for \eqn{Y \sim M} p-values.}

\item{idx_Xgene}{unique index for X gene}

\item{idx_Mediator}{unique index for Mediator}

\item{idx_Ygene}{unique index for Y gene}

\item{idx_Ysite}{unique index for Y site}

\item{Est_param_YX}{Estimated regression coefficient of \eqn{X} in the linear model \eqn{Y \sim M + X}.}

\item{Est_param_YM}{Estimated regression coefficient of \eqn{M} in the linear model \eqn{Y \sim M + X}.}

\item{Est_param_MX}{Estimated regression coefficient of \eqn{X} in the linear model \eqn{M \sim X}.}

\item{numCores}{integer, number of cores for parallel backend}

\item{y_name}{character, name of the outcome in Y_list}
}
\value{
A data frame of row-bound results across matching keys.
}
\description{
Integrative pathway-based mediation analysis (iProPath
}
