% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdact_sample_parallel.R
\name{mdact_sample_parallel}
\alias{mdact_sample_parallel}
\title{Parallel mDACT sampling-based test}
\usage{
mdact_sample_parallel(p.M, p.M.uni, p.Y, pi1_MX, pi1_YM, numCores, y_name)
}
\arguments{
\item{p.M}{numeric vector of p-values from M ~ X regressions}

\item{p.M.uni}{unique numeric vector of p-values from M ~ X regressions}

\item{p.Y}{numeric vector of p-values from Y ~ M + X  regressions}

\item{pi1_MX}{estimated proportion of non-null signals for M ~ X p-values.}

\item{pi1_YM}{estimated proportion of non-null signals for Y ~ M p-values.}

\item{numCores}{integer, number of cores for parallel backend}

\item{y_name}{character, name of the outcome in Y_list}
}
\value{
A list with the following components:
\describe{
\item{MDACT}{Numeric vector of MDACT results.}
\item{pi_df}{A data frame containing estimated \eqn{\pi} values.}
}
}
\description{
Parallel mDACT sampling-based test
}
