% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnullPropEst.R
\name{nonnullPropEst_JC}
\alias{nonnullPropEst_JC}
\title{Estimate non-null proportion (JC method)}
\usage{
nonnullPropEst_JC(p, u = 0, sigma = 1, parallel = F)
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{u}{Numeric; tuning parameter (default 0).}

\item{sigma}{Numeric; noise scale parameter (default 1).}

\item{parallel}{Logical; whether to run computation in parallel.}
}
\value{
epsest Estimated non-null proportion.
}
\description{
Estimate non-null proportion (JC method)
}
