% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnullPropEst.R
\name{nonnullPropEst_locfdr}
\alias{nonnullPropEst_locfdr}
\title{Estimate non-null proportion (local FDR)}
\usage{
nonnullPropEst_locfdr(p, est = c("central matching", "mle"), u = 0, sigma = 1)
}
\arguments{
\item{p}{Numeric vector of p-values.}

\item{est}{Character string specifying the locfdr estimation method;
either \code{"central matching"} or \code{"mle"}.}

\item{u}{Numeric value specifying the mean of the null distribution.}

\item{sigma}{Numeric value specifying the standard deviation of the null distribution.}
}
\value{
Numeric scalar giving the estimated non-null proportion.
}
\description{
Estimate non-null proportion (local FDR)
}
