% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonnullPropEst.R
\name{nonnullPropEst_q}
\alias{nonnullPropEst_q}
\title{Estimate non-null proportion (q-value based)}
\usage{
nonnullPropEst_q(x, pi0.method = c("smoother", "bootstrap"))
}
\arguments{
\item{x}{Numeric vector of p-values.}

\item{pi0.method}{Character string specifying the \eqn{\pi_0} estimation method;
either \code{"smoother"} or \code{"bootstrap"}.}
}
\value{
Numeric scalar giving the estimated non-null proportion.
}
\description{
Estimate non-null proportion (q-value based)
}
