% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ipropath.R
\name{run_ipropath}
\alias{run_ipropath}
\title{Run iProPath core algorithm}
\usage{
run_ipropath(merged_df, pi1_MX, pi1_YM, num_cores = 1L, y_name)
}
\arguments{
\item{merged_df}{Data frame containing columns required by \code{ipropath()}:
p_value_YX, p_value_MX, p_value_YM, idx_Xgene, idx_Mediator, idx_Ygene, idx_Ysite,
Est_param_YX, Est_param_YM, Est_param_MX.}

\item{pi1_MX}{Estimated non-null proportion for M~X.}

\item{pi1_YM}{Estimated non-null proportion for Y~M.}

\item{num_cores}{Integer number of CPU cores to use.}

\item{y_name}{Character string passed to \code{ipropath()}.}
}
\value{
A list returned by \code{ipropath()}, with an additional element
\code{runtime_secs} giving the elapsed time in seconds.
}
\description{
Run iProPath core algorithm
}
