% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ipropath_results.R
\name{summarize_ipropath_results}
\alias{summarize_ipropath_results}
\title{Summarize iProPath results and write to file}
\usage{
summarize_ipropath_results(
  res_all,
  real_data.pi_df = NULL,
  ipropath.pi_df,
  file_name,
  output_dir = "Results"
)
}
\arguments{
\item{res_all}{...}

\item{real_data.pi_df}{...}

\item{ipropath.pi_df}{...}

\item{file_name}{...}

\item{output_dir}{...}
}
\value{
NULL. Results are written to a text file.
}
\description{
Summarize iProPath results and write to file
}
