% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropindex_and_transpose}
\alias{dropindex_and_transpose}
\title{Drop index columns and transpose data frame}
\usage{
dropindex_and_transpose(df)
}
\arguments{
\item{df}{data frame}
}
\value{
transposed matrix
}
\description{
Drop index columns and transpose data frame
}
